+++ src/components/archive/header.vue
@@ -65,7 +65,6 @@
 
   .header {
     position: relative;
-    overflow: hidden;
     display: flex;
     flex-direction: column;
     width: 100%;